tests/e2e/test_authentication_flow.py
"""
E2E Tests for Authentication Flow
"""

import pytest
import jwt
import time
from datetime import datetime, timedelta
from httpx import AsyncClient


class TestAuthenticationFlow:
    """Test suite for complete authentication user journey"""
    
    @pytest.mark.asyncio
    async def test_user_registration_flow(self, client: AsyncClient):
        """Test complete user registration flow"""
        # Step 1: Register new user
        user_data = {
            "email": "newuser@example.com",
            "username": "newuser",
            "password": "SecurePassword123!",
            "full_name": "New User"
        }
        
        response = await client.post("/api/v1/auth/register", json=user_data)
        
        # TODO: Implement registration endpoint in api/main.py
        # assert response.status_code == 201
        # data = response.json()
        # assert "id" in data
        # assert data["email"] == user_data["email"]
        # assert data["username"] == user_data["username"]
        # assert "hashed_password" not in data  # Password should not be returned
    
    @pytest.mark.asyncio
    async def test_user_login_flow(self, client: AsyncClient):
        """Test complete user login flow"""
        # Step 1: Register user first
        user_data = {
            "email": "logintest@example.com",
            "username": "logintest",
            "password": "LoginPassword123!",
            "full_name": "Login Test User"
        }
        
        # Register user
        # TODO: Uncomment when registration endpoint is implemented
        # await client.post("/api/v1/auth/register", json=user_data)
        
        # Step 2: Login with credentials
        login_data = {
            "username": user_data["email"],  # Can use email or username
            "password": user_data["password"]
        }
        
        # TODO: Implement login endpoint in api/main.py
        # response = await client.post("/api/v1/auth/login", data=login_data)
        # assert response.status_code == 200
        # 
        # token_data = response.json()
        # assert "access_token" in token_data
        # assert "token_type" in token_data
        # assert token_data["token_type"] == "bearer"
        # assert "refresh_token" in token_data
    
    @pytest.mark.asyncio
    async def test_jwt_token_validation(self, client: AsyncClient):
        """Test JWT token structure and claims"""
        # TODO: Implement when auth endpoints are ready
        pass
        # # Create and login user
        # user_data = {
        #     "email": "jwttest@example.com",
        #     "username": "jwttest",
        #     "password": "JWTPassword123!",
        #     "full_name": "JWT Test User"
        # }
        # 
        # await client.post("/api/v1/auth/register", json=user_data)
        # 
        # login_response = await client.post(
        #     "/api/v1/auth/login",
        #     data={
        #         "username": user_data["email"],
        #         "password": user_data["password"]
        #     }
        # )
        # 
        # token = login_response.json()["access_token"]
        # 
        # # Decode token (without verification for testing structure)
        # decoded = jwt.decode(token, options={"verify_signature": False})
        # 
        # assert "sub" in decoded  # Subject (user identifier)
        # assert "exp" in decoded  # Expiration time
        # assert "iat" in decoded  # Issued at time
        # 
        # # Check expiration is set correctly (30 minutes from now)
        # exp_time = datetime.fromtimestamp(decoded["exp"])
        # expected_exp = datetime.utcnow() + timedelta(minutes=30)
        # assert abs((exp_time - expected_exp).total_seconds()) < 60  # Within 1 minute tolerance
    
    @pytest.mark.asyncio
    async def test_protected_endpoint_access(self, client: AsyncClient):
        """Test accessing protected endpoints with and without authentication"""
        # TODO: Implement when auth endpoints are ready
        pass
        # # Try to access protected endpoint without auth
        # response = await client.get("/api/v1/users/me")
        # assert response.status_code == 401
        # 
        # # Create authenticated client
        # auth_client = await authenticated_client(client)
        # 
        # # Access protected endpoint with auth
        # response = await auth_client.get("/api/v1/users/me")
        # assert response.status_code == 200
        # 
        # user_data = response.json()
        # assert "email" in user_data
        # assert "username" in user_data
    
    @pytest.mark.asyncio
    async def test_token_refresh_flow(self, client: AsyncClient):
        """Test token refresh mechanism"""
        # TODO: Implement when auth endpoints are ready
        pass
        # # Login to get tokens
        # login_response = await client.post(
        #     "/api/v1/auth/login",
        #     data={
        #         "username": "test@example.com",
        #         "password": "TestPassword123!"
        #     }
        # )
        # 
        # tokens = login_response.json()
        # refresh_token = tokens["refresh_token"]
        # 
        # # Use refresh token to get new access token
        # refresh_response = await client.post(
        #     "/api/v1/auth/refresh",
        #     json={"refresh_token": refresh_token}
        # )
        # 
        # assert refresh_response.status_code == 200
        # new_tokens = refresh_response.json()
        # assert "access_token" in new_tokens
        # assert new_tokens["access_token"] != tokens["access_token"]
    
    @pytest.mark.asyncio
    async def test_logout_flow(self, client: AsyncClient):
        """Test user logout flow"""
        # TODO: Implement when auth endpoints are ready
        pass
        # # Create authenticated client
        # auth_client = await authenticated_client(client)
        # 
        # # Logout
        # response = await auth_client.post("/api/v1/auth/logout")
        # assert response.status_code == 200
        # 
        # # Try to access protected endpoint after logout
        # response = await auth_client.get("/api/v1/users/me")
        # assert response.status_code == 401
    
    @pytest.mark.asyncio
    async def test_registration_validation(self, client: AsyncClient):
        """Test registration input validation"""
        # Test invalid email
        invalid_data = {
            "email": "invalid-email",
            "username": "testuser",
            "password": "Password123!",
            "full_name": "Test User"
        }
        
        # TODO: Implement validation in registration endpoint
        # response = await client.post("/api/v1/auth/register", json=invalid_data)
        # assert response.status_code == 422
        # assert "email" in response.json()["detail"][0]["loc"]
        
        # Test weak password
        weak_password_data = {
            "email": "test@example.com",
            "username": "testuser",
            "password": "weak",
            "full_name": "Test User"
        }
        
        # response = await client.post("/api/v1/auth/register", json=weak_password_data)
        # assert response.status_code == 422
        # assert "password" in str(response.json()["detail"])
    
    @pytest.mark.asyncio
    async def test_duplicate_registration(self, client: AsyncClient):
        """Test that duplicate registrations are prevented"""
        user_data = {
            "email": "duplicate@example.com",
            "username": "duplicate",
            "password": "Password123!",
            "full_name": "Duplicate User"
        }
        
        # TODO: Implement when registration endpoint is ready
        # # First registration should succeed
        # response1 = await client.post("/api/v1/auth/register", json=user_data)
        # assert response1.status_code == 201
        # 
        # # Second registration with same email should fail
        # response2 = await client.post("/api/v1/auth/register", json=user_data)
        # assert response2.status_code == 400
        # assert "already exists" in response2.json()["detail"].lower()
    
    @pytest.mark.asyncio
    async def test_invalid_login_credentials(self, client: AsyncClient):
        """Test login with invalid credentials"""
        # TODO: Implement when login endpoint is ready
        pass
        # # Try to login with non-existent user
        # response = await client.post(
        #     "/api/v1/auth/login",
        #     data={
        #         "username": "nonexistent@example.com",
        #         "password": "SomePassword123!"
        #     }
        # )
        # assert response.status_code == 401
        # assert "incorrect" in response.json()["detail"].lower()
        # 
        # # Try to login with wrong password
        # response = await client.post(
        #     "/api/v1/auth/login",
        #     data={
        #         "username": "test@example.com",
        #         "password": "WrongPassword123!"
        #     }
        # )
        # assert response.status_code == 401
    
    @pytest.mark.asyncio
    async def test_auth_performance(self, client: AsyncClient, performance_threshold):
        """Test authentication endpoints performance"""
        # TODO: Implement when auth endpoints are ready
        pass
        # import time
        # 
        # # Test registration performance
        # start = time.time()
        # await client.post("/api/v1/auth/register", json={
        #     "email": f"perf{time.time()}@example.com",
        #     "username": f"perf{time.time()}",
        #     "password": "Password123!",
        #     "full_name": "Performance Test"
        # })
        # registration_time = time.time() - start
        # assert registration_time < performance_threshold["api_response_time"]
        # 
        # # Test login performance
        # start = time.time()
        # await client.post("/api/v1/auth/login", data={
        #     "username": "test@example.com",
        #     "password": "TestPassword123!"
        # })
        # login_time = time.time() - start
        # assert login_time < performance_threshold["api_response_time"]