requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis
redis==5.0.1
hiredis==2.2.3

# File Processing
python-magic==0.4.27
pypdf==3.17.1
python-docx==1.1.0
pillow==10.1.0

# Claude API
anthropic==0.8.0
tiktoken==0.5.2

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Rate Limiting
slowapi==0.1.9

# Background Tasks
celery==5.3.4
redis==5.0.1

# MCP Server
mcp==0.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.2